requirements.txt.txt
Flask
Flasgger
mysql-connector-python
gunicorn
Werkzeug
